Move port finder closer to aspect
The only place where it's used.
